+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/hardware/M5ExtraLargeEphemeral.kt
@@ -5,18 +5,19 @@ import com.atlassian.performance.tools.awsinfrastructure.hardware.EphemeralStora
 import com.atlassian.performance.tools.ssh.api.SshConnection
 
 /**
- * C5, 9 XL, ephemeral storage.
+ * M5, XL, ephemeral storage.
  *
  * See [AWS docs](https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/InstanceStorage.html).
  * > The data in an instance store persists only during the lifetime of its associated instance.
  * > (...) Do not rely on instance store for valuable, long-term data.
  * If you need your data to survive long periods of time, use a durable computer.
  *
- * @see C4EightExtraLargeElastic
+ * @since 2.7.0
+ * @see M4ExtraLargeElastic
  */
-class C5NineExtraLargeEphemeral : Computer() {
+class M5ExtraLargeEphemeral : Computer() {
 
-    override val instanceType: InstanceType = InstanceType.C5d9xlarge
+    override val instanceType: InstanceType = InstanceType.M5dXlarge
 
     override fun setUp(ssh: SshConnection) {
         EphemeralStorage().mount(ssh)